fix(gateway): optional heartbeat interval
There was a bug: if the client was not connected to the socket for a long time, then after connecting the client would flood `ClientEvent::Ping` packets, and the number of packets was equal to the waiting time divided by the heartbeat period. For example, in just 10 minutes of inactivity, about 40 packets will be sent simultaneously.... meh.

I decided to fix this by making `heartbeat_interval` `None` when the client disconnects and when it connects AND receives the `ServerEvent::Authenticated` packet - `Some(...)`.

I haven’t tested it yet though :)

Well, I’ll put the bot on overnight and check its behavior.